First pass at making goa Windows compatible.
Need to merge to master so `go get ...goagen` picks up the changes for
appveyor Windows build.
