+++ appveyor.yml
@@ -1,0 +1,23 @@
+version: "{build}"
+
+os: Windows Server 2012 R2
+
+environment:
+  GOPATH: c:\gopath
+
+install:
+  - go version
+  - go env
+  - mkdir c:\gopath
+  - copy c:\MinGW\bin\mingw32-make.exe c:\MinGW\bin\make.exe
+  - set PATH=%PATH%;c:\MinGW\bin;%GOPATH%\bin
+  - go get github.com/raphael/goa/goagen
+
+build_script:
+  - make depend
+  - make goagen
+  - make test
+
+test: off
+
+deploy: off
