+++ goagen/utils/signal.go
@@ -3,17 +3,7 @@ package utils
 import (
 	"os"
 	"os/signal"
-	"syscall"
 )
-
-var defaultSignals = []os.Signal{
-	syscall.SIGHUP,
-	syscall.SIGINT,
-	syscall.SIGQUIT,
-	syscall.SIGTERM,
-	syscall.SIGUSR1,
-	syscall.SIGUSR2,
-}
 
 // Catch signals and invoke then callback
 func Catch(signals []os.Signal, then func()) {
