+++ goagen/utils/signal_lin.go
@@ -1,0 +1,17 @@
+// +build !windows
+
+package utils
+
+import (
+	"os"
+	"syscall"
+)
+
+var defaultSignals = []os.Signal{
+	syscall.SIGHUP,
+	syscall.SIGINT,
+	syscall.SIGQUIT,
+	syscall.SIGTERM,
+	syscall.SIGUSR1,
+	syscall.SIGUSR2,
+}
