+++ goagen/utils/signal_win.go
@@ -1,0 +1,15 @@
+// +build windows
+
+package utils
+
+import (
+	"os"
+	"syscall"
+)
+
+var defaultSignals = []os.Signal{
+	syscall.SIGHUP,
+	syscall.SIGINT,
+	syscall.SIGQUIT,
+	syscall.SIGTERM,
+}
